pydoc: replace goconvey -> testify for unit tests
There's not a ton of tests here and they're very straightforward,
so I converted them to use `stretchr/testify` instead of `goconvey`
to avoid pulling in another dependency/having tests in wildly
disparate styles.

These tests came from `luci-go` project along with the parser, but
it's been stable for 3+ years, so it's unlikely that this change
will cause a headache down the road if we need to pull in more
changes from upstream.