analysis: add global/built-in function support (#1)
The `Analyzer` is a semi-stateful type to handle analysis for a
workspace. Currently, the only "state" it has is built-in functions
and symbols, which can be registered with it at construction.

For generic CLI usage, the `--builtin-paths` can be used to pass
Starlark/Python3 (including type hints) stub files, which will be
parsed at startup and treated as globals.

In the future, the `Analyzer` can/should handle cross-document
analysis/resolutions as well. This is in contrast to the purely
functional analysis methods that only consider a local Tree-sitter
(sub)tree.