analysis: determine active param
Count `,` symbols within the arg nodes to determine which parameter
is active.

This won't handle pathological cases, e.g. `foo(x,,,,,)` will end
up grouping some of the `,`s into a single anonymous node, so we
won't see the active param as having "advanced". This seems
consistent with the VSCode Python extension behavior, so is not
really worth addressing at the moment (or ever?)