document: add leak checker to server tests (#9)
It's easy to leak Tree-sitter tree objects by forgetting to call
`Close()`. This is typically done by calling `Close()` on the
wrapping `Document` object.

To prevent this, `Document` is now an interface rather than a
struct, and a special implementation is used in tests that wraps
the real one but does some basic ref counting to ensure that they
are being properly closed and will fail a test if any are left
over at the end. (This does means tests need to remember to clean
up any docs they read manually and/or use helpers that do it for
them, e.g. `requireDocContents`.)